feat(layout): A few improvements to the existing layout
