+++ src/components/header/Header.tsx
@@ -1,0 +1,31 @@
+import Image from "next/image";
+
+import Wrapper from '../utility/Wrapper';
+import CartPreview from './CartPreview';
+
+function Header() {
+    const styles = [
+        "sticky top-0 z-50", // position
+        "h-24", // sizing
+        "bg-white border-b border-slate", // theme
+    ];
+    return (
+        <header className={styles.join(' ')}>
+            <Wrapper className="py-2 flex items-center">
+                <div className="grow">
+                    <Image
+                        src="/FakeShop.svg"
+                        alt="FakeShop Logo"
+                        className="dark:invert"
+                        width={300}
+                        height={80}
+                        priority
+                    />
+                </div>
+                <CartPreview />
+            </Wrapper>        
+        </header>
+    );
+}
+
+export default Header;
